+++ LICENSE.md
@@ -1,6 +1,6 @@
 The MIT License (MIT)
 
-Copyright (c) Tighten Co. <hello@tighten.co>
+Copyright (c) Jordan Partridge <jordan@partridge.rocks>
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
