+++ README.md
@@ -1,18 +1,22 @@
-# :package_fullname
+# LaraHub
 
-[![Latest Version on Packagist](https://img.shields.io/packagist/v/tightenco/:package_name.svg?style=flat-square)](https://packagist.org/packages/tightenco/:package_name)
-[![GitHub Tests Action Status](https://img.shields.io/github/actions/workflow/status/tighten/:package_name/.github/workflows/tests.yml?branch=main&label=tests)](https://github.com/tighten/:package_name/actions?query=workflow%3Arun-tests+branch%3Amain)
+[![Latest Version on Packagist](https://img.shields.io/packagist/v/tightenco/larahub.svg?style=flat-square)](https://packagist.org/packages/tightenco/larahub)
+[![GitHub Tests Action Status](https://img.shields.io/github/actions/workflow/status/tighten/larahub/.github/workflows/tests.yml?branch=main&label=tests)](https://github.com/tighten/larahub/actions?query=workflow%3Arun-tests+branch%3Amain)
 
-**Note:** Replace ```:author_name``` ```:author_username``` ```:author_email``` ```:package_name``` ```:package_description``` ```:package_php_namespace``` with their correct values in [README.md](README.md), [composer.json](composer.json), and all three files under `src/`, then delete this line.
 
-:package_description
+
+Bridging the Gap betwen Laravel and Github
 
 ## Installation
+
+
+
+
 
 You can install the package via composer:
 
 ```bash
-composer require tightenco/:package_name
+composer require partridge-rocks/larahub
 ```
 
 ## Usage
@@ -37,7 +41,7 @@ If you discover any security related issues, please email hello@tighten.co inste
 
 ## Credits
 
-- [:author_name](https://github.com/:author_username)
+- [Jordan Partridge](https://github.com/jordanpartridge)
 - [All Contributors](../../contributors)
 
 ## License
