--- src/Skeleton.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Tighten\:package_php_namespace;
-
-class :package_php_namespace
-{
-    public function __construct()
-    {
-    }
-}
