--- src/SkeletonFacade.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace Tighten\:package_php_namespace;
-
-use Illuminate\Support\Facades\Facade;
-
-class :package_php_namespaceFacade extends Facade
-{
-    protected static function getFacadeAccessor()
-    {
-        return ':package_name';
-    }
-}
