+++ src/partridge-rocks.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Tighten\partridge-rocks;
+
+class partridge-rocks
+{
+    public function __construct()
+    {
+    }
+}
